backend/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
numpy==1.25.2
psycopg2==2.9.6
pytz==2023.3
scipy==1.11.2
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3